+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 - Changed the initial value of default material conversion settings to Toon Standard. (VRCSDK 3.8.1 or later)
+- Improved the appearance when converting lilToon normal mode MatCap to Toon Standard.
 
 ## [2.11.0] - 2025-06-22